Move level validation to separate file
