+++ cedar-policy-validator/src/lib.rs
@@ -31,13 +31,11 @@
 #![allow(clippy::result_large_err, clippy::large_enum_variant)] // see #878
 #![cfg_attr(feature = "wasm", allow(non_snake_case))]
 
-#[cfg(feature = "level-validate")]
-use cedar_policy_core::ast::{BinaryOp, PolicyID};
 use cedar_policy_core::ast::{Policy, PolicySet, Template};
 use serde::Serialize;
 use std::collections::HashSet;
 #[cfg(feature = "level-validate")]
-use validation_errors::{EntityDerefLevel, EntityDerefLevelViolation};
+mod level_validate;
 
 #[cfg(feature = "entity-manifest")]
 pub mod entity_manifest;
@@ -59,8 +57,6 @@ mod str_checks;
 pub use str_checks::confusable_string_checks;
 pub mod cedar_schema;
 pub mod typecheck;
-#[cfg(feature = "level-validate")]
-use typecheck::PolicyCheck;
 use typecheck::Typechecker;
 
 pub mod types;
@@ -195,33 +191,6 @@ impl Validator {
         (validation_errors.chain(errors), warnings)
     }
 
-    #[cfg(feature = "level-validate")]
-    /// Run `validate_policy` in strict mode against a single static policy or template (note
-    /// that Core `Template` includes static policies as well), gathering all
-    /// validation errors and warnings in the returned iterators.
-    /// If strict validation passes, we will also perform level validation (see RFC 76).
-    fn strict_validate_policy_with_level<'a>(
-        &'a self,
-        p: &'a Template,
-        max_deref_level: u32,
-    ) -> (
-        impl Iterator<Item = ValidationError> + 'a,
-        impl Iterator<Item = ValidationWarning> + 'a,
-    ) {
-        let (errors, warnings) = self.validate_policy(p, ValidationMode::Strict);
-
-        let mut peekable_errors = errors.peekable();
-
-        // Only perform level validation if strict validation passed.
-        if peekable_errors.peek().is_none() {
-            let levels_errors =
-                self.check_entity_deref_level(p, &EntityDerefLevel::from(max_deref_level), p.id());
-            (peekable_errors.chain(levels_errors), warnings)
-        } else {
-            (peekable_errors.into_iter().chain(vec![]), warnings)
-        }
-    }
-
     /// Run relevant validations against a single template-linked policy,
     /// gathering all validation errors together in the returned iterator.
     fn validate_slots<'a>(
@@ -266,205 +235,6 @@ impl Validator {
         let mut warnings = HashSet::new();
         typecheck.typecheck_policy(t, &mut errors, &mut warnings);
         (errors.into_iter(), warnings.into_iter())
-    }
-
-    #[cfg(feature = "level-validate")]
-    /// Check that `t` respects `max_allowed_level`
-    /// This assumes that (strict) typechecking has passed
-    fn check_entity_deref_level<'a>(
-        &'a self,
-        t: &'a Template,
-        max_allowed_level: &EntityDerefLevel,
-        policy_id: &PolicyID,
-    ) -> Vec<ValidationError> {
-        let typechecker = Typechecker::new(&self.schema, ValidationMode::Strict, t.id().clone());
-        let type_annotated_asts = typechecker.typecheck_by_request_env(t);
-        let mut errs = vec![];
-        for (_, policy_check) in type_annotated_asts {
-            match policy_check {
-                PolicyCheck::Success(e) | PolicyCheck::Irrelevant(_, e) => {
-                    let res =
-                        self.check_entity_deref_level_helper(&e, max_allowed_level, policy_id);
-                    match res.1 {
-                        Some(e) => errs.push(ValidationError::EntityDerefLevelViolation(e)),
-                        None => (),
-                    }
-                }
-                // PANIC SAFETY: We only validate the level after strict validation passed
-                #[allow(clippy::unreachable)]
-                PolicyCheck::Fail(_) => unreachable!(),
-            }
-        }
-        errs
-    }
-
-    #[cfg(feature = "level-validate")]
-    fn min(
-        v: impl IntoIterator<Item = (EntityDerefLevel, Option<EntityDerefLevelViolation>)>,
-    ) -> (EntityDerefLevel, Option<EntityDerefLevelViolation>) {
-        let p = v.into_iter().min_by(|(l1, _), (l2, _)| l1.cmp(l2));
-        match p {
-            Some(p) => p.clone(),
-            None => (EntityDerefLevel { level: 0 }, None),
-        }
-    }
-
-    #[cfg(feature = "level-validate")]
-    /// Walk the type-annotated AST and compute the used level and possible violation
-    /// Returns a tuple of `(actual level used, optional violation information)`
-    fn check_entity_deref_level_helper<'a>(
-        &'a self,
-        e: &cedar_policy_core::ast::Expr<Option<crate::types::Type>>,
-        max_allowed_level: &EntityDerefLevel,
-        policy_id: &PolicyID,
-    ) -> (EntityDerefLevel, Option<EntityDerefLevelViolation>) {
-        use crate::types::{EntityRecordKind, Type};
-        use cedar_policy_core::ast::ExprKind;
-        match e.expr_kind() {
-            ExprKind::Lit(_) => (
-                EntityDerefLevel { level: 0 }, //Literals can't be dereferenced
-                None,
-            ),
-            ExprKind::Var(_) => (max_allowed_level.clone(), None), //Roots start at `max_allowed_level`
-            ExprKind::Slot(_) => (EntityDerefLevel { level: 0 }, None), //Slot will be replaced by Entity literal so treat the same
-            ExprKind::Unknown(_) => (
-                EntityDerefLevel { level: 0 }, //Can't dereference an unknown
-                None,
-            ),
-            ExprKind::If {
-                test_expr,
-                then_expr,
-                else_expr,
-            } => {
-                let es = [test_expr, then_expr, else_expr];
-                let v: Vec<(EntityDerefLevel, Option<_>)> = es
-                    .iter()
-                    .map(|l| self.check_entity_deref_level_helper(l, max_allowed_level, policy_id))
-                    .collect();
-                Self::min(v)
-            }
-            ExprKind::And { left, right } | ExprKind::Or { left, right } => {
-                let es = [left, right];
-                let v: Vec<(EntityDerefLevel, Option<_>)> = es
-                    .iter()
-                    .map(|l| self.check_entity_deref_level_helper(l, max_allowed_level, policy_id))
-                    .collect();
-                Self::min(v)
-            }
-            ExprKind::UnaryApp { arg, .. } => {
-                self.check_entity_deref_level_helper(arg, max_allowed_level, policy_id)
-            }
-            // `In` operator decrements the LHS only
-            ExprKind::BinaryApp { op, arg1, arg2 } if op == &BinaryOp::In => {
-                let mut lhs =
-                    self.check_entity_deref_level_helper(arg1, max_allowed_level, policy_id);
-                let rhs = self.check_entity_deref_level_helper(arg2, max_allowed_level, policy_id);
-                lhs = (lhs.0.decrement(), lhs.1);
-                let new_level = Self::min(vec![lhs, rhs]).0;
-                if new_level.level < 0 {
-                    (
-                        new_level,
-                        Some(EntityDerefLevelViolation {
-                            source_loc: e.source_loc().cloned(),
-                            policy_id: policy_id.clone(),
-                            actual_level: new_level,
-                            allowed_level: max_allowed_level.clone(),
-                        }),
-                    )
-                } else {
-                    (new_level, None)
-                }
-            }
-            ExprKind::BinaryApp { arg1, arg2, .. } => {
-                let es = [arg1, arg2];
-                let v: Vec<(EntityDerefLevel, Option<_>)> = es
-                    .iter()
-                    .map(|l| self.check_entity_deref_level_helper(l, max_allowed_level, policy_id))
-                    .collect();
-                Self::min(v)
-            }
-            ExprKind::ExtensionFunctionApp { args, .. } => {
-                let v: Vec<(EntityDerefLevel, Option<_>)> = args
-                    .iter()
-                    .map(|l| self.check_entity_deref_level_helper(l, max_allowed_level, policy_id))
-                    .collect();
-                Self::min(v)
-            }
-            ExprKind::GetAttr { expr, attr }
-                if matches!(expr.expr_kind(), ExprKind::Record(..)) =>
-            {
-                match expr.expr_kind() {
-                    ExprKind::Record(m) => {
-                        // PANIC SAFETY: Strict validation checked that this access is safe
-                        #[allow(clippy::unwrap_used)]
-                        self.check_entity_deref_level_helper(
-                            m.get(attr).unwrap(),
-                            max_allowed_level,
-                            policy_id,
-                        )
-                    }
-                    // PANIC SAFETY: We just checked that this node is a Record
-                    #[allow(clippy::unreachable)]
-                    _ => unreachable!(),
-                }
-            }
-            ExprKind::GetAttr { expr, .. } | ExprKind::HasAttr { expr, .. } => match expr
-                .as_ref()
-                .data()
-            {
-                Some(ty) => {
-                    let child_level_info =
-                        self.check_entity_deref_level_helper(expr, max_allowed_level, policy_id);
-                    match ty {
-                        Type::EntityOrRecord(EntityRecordKind::Entity { .. })
-                        | Type::EntityOrRecord(EntityRecordKind::ActionEntity { .. }) => {
-                            let child_level = child_level_info.0;
-                            let new_level = child_level.decrement();
-                            if new_level.level < 0 {
-                                (
-                                    new_level,
-                                    Some(EntityDerefLevelViolation {
-                                        source_loc: e.source_loc().cloned(),
-                                        policy_id: policy_id.clone(),
-                                        actual_level: new_level,
-                                        allowed_level: max_allowed_level.clone(),
-                                    }),
-                                )
-                            } else {
-                                (new_level, None)
-                            }
-                        }
-                        Type::EntityOrRecord(EntityRecordKind::AnyEntity) => {
-                            // AnyEntity cannot be dereferenced
-                            (EntityDerefLevel { level: 0 }, None)
-                        }
-                        _ => child_level_info,
-                    }
-                }
-                // PANIC SAFETY: Strict validation passed, so annotating the AST will succeed
-                #[allow(clippy::unreachable)]
-                None => unreachable!("Expected type-annotated AST"),
-            },
-            ExprKind::Like { expr, .. } | ExprKind::Is { expr, .. } => {
-                self.check_entity_deref_level_helper(expr, max_allowed_level, policy_id)
-            }
-            ExprKind::Set(elems) => {
-                let v: Vec<(EntityDerefLevel, Option<_>)> = elems
-                    .iter()
-                    .map(|l| self.check_entity_deref_level_helper(l, max_allowed_level, policy_id))
-                    .collect();
-                Self::min(v)
-            }
-            ExprKind::Record(fields) => {
-                let v: Vec<(EntityDerefLevel, Option<_>)> = fields
-                    .iter()
-                    .map(|(_, l)| {
-                        self.check_entity_deref_level_helper(l, max_allowed_level, policy_id)
-                    })
-                    .collect();
-                Self::min(v)
-            }
-        }
     }
 }
 
@@ -769,89 +539,5 @@ mod test {
                 "һenry"
             )]
         );
-    }
-}
-
-#[cfg(feature = "level-validate")]
-#[cfg(test)]
-mod levels_validation_tests {
-    use super::*;
-    use cedar_policy_core::parser;
-
-    fn get_schema() -> ValidatorSchema {
-        json_schema::Fragment::from_json_str(
-            r#"
-            {
-                "": {
-                    "entityTypes": {
-                        "User": {
-                            "memberOfTypes": ["User"]
-                        },
-                        "Photo": {
-                            "shape": {
-                                "type": "Record",
-                                "attributes": {
-                                    "foo": {
-                                        "type": "Entity",
-                                        "name": "User",
-                                        "required": true
-                                    }
-                                }
-                            }
-                        }
-                    },
-                    "actions": {
-                        "view": {
-                            "appliesTo": {
-                                "resourceTypes": [ "Photo" ],
-                                "principalTypes": [ "User" ]
-                            }
-                        }
-                    }
-                }
-            }
-        "#,
-        )
-        .expect("Schema parse error.")
-        .try_into()
-        .expect("Expected valid schema.")
-    }
-
-    #[test]
-    fn test_levels_validation_passes() {
-        let schema = get_schema();
-        let validator = Validator::new(schema);
-
-        let mut set = PolicySet::new();
-        let src = r#"permit(principal == User::"һenry", action, resource) when {1 > 0};"#;
-        let p = parser::parse_policy(None, src).unwrap();
-        set.add_static(p).unwrap();
-
-        let template_name = PolicyID::from_string("policy0");
-        let result = validator.check_entity_deref_level(
-            set.get_template(&template_name).unwrap(),
-            &EntityDerefLevel { level: 0 },
-            &template_name,
-        );
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_levels_validation_fails() {
-        let schema = get_schema();
-        let validator = Validator::new(schema);
-
-        let mut set = PolicySet::new();
-        let src = r#"permit(principal == User::"һenry", action, resource) when {principal in resource.foo};"#;
-        let p = parser::parse_policy(None, src).unwrap();
-        set.add_static(p).unwrap();
-
-        let template_name = PolicyID::from_string("policy0");
-        let result = validator.check_entity_deref_level(
-            set.get_template(&template_name).unwrap(),
-            &EntityDerefLevel { level: 0 },
-            &template_name,
-        );
-        assert!(result.len() == 1);
     }
 }
